Optical Sensor (sentinel) compatibility
Tweak the code a bit in order for filament_detector module to support
optical sensor. The switch.filamentout module seemed to detect "pin
fall" event in order to pause print, which allowed to restart a print
even if no filament was detected.

The filament detector periodically check the status of the optical_pin,
meaning it will return to "out of filament" state if the print was
restarted without changing the filament.
detection_type was added to the config to specify which mode to use (1
for encoder, 2 for optical sensor, default is 1)